+++ src/gafaelfawr/handlers/api.py
@@ -21,6 +21,7 @@ from fastapi import (
     Response,
     status,
 )
+from safir.models import ErrorModel
 
 from gafaelfawr.constants import ACTOR_REGEX, CURSOR_REGEX, USERNAME_REGEX
 from gafaelfawr.dependencies.auth import AuthenticateRead, AuthenticateWrite
@@ -28,7 +29,6 @@ from gafaelfawr.dependencies.context import RequestContext, context_dependency
 from gafaelfawr.exceptions import ErrorLocation, NotFoundError
 from gafaelfawr.models.admin import Admin
 from gafaelfawr.models.auth import APIConfig, APILoginResponse, Scope
-from gafaelfawr.models.error import ErrorModel
 from gafaelfawr.models.history import TokenChangeHistoryEntry
 from gafaelfawr.models.token import (
     AdminTokenRequest,