+++ src/gafaelfawr/middleware/state.py
@@ -28,7 +28,7 @@ class BaseState(ABC):
 
     @classmethod
     @abstractmethod
-    def from_cookie(cls, cookie: str, request: Request) -> BaseState:
+    async def from_cookie(cls, cookie: str, request: Request) -> BaseState:
         """Reconstruct state from an encrypted cookie.
 
         Parameters
@@ -45,7 +45,7 @@ class BaseState(ABC):
         """
 
     @abstractmethod
-    def as_cookie(self) -> str:
+    async def as_cookie(self) -> str:
         """Build an encrypted cookie representation of the state.
 
         Returns
@@ -97,7 +97,7 @@ class StateMiddleware(BaseHTTPMiddleware):
     ) -> Response:
         if self.cookie_name in request.cookies:
             cookie = request.cookies[self.cookie_name]
-            state = self.state_class.from_cookie(cookie, request)
+            state = await self.state_class.from_cookie(cookie, request)
         else:
             state = self.state_class()
 
@@ -111,7 +111,7 @@ class StateMiddleware(BaseHTTPMiddleware):
 
         # If the state has changed, write out the new state.
         if request.state.cookie != state:
-            cookie = request.state.cookie.as_cookie()
+            cookie = await request.state.cookie.as_cookie()
             secure = self.is_cookie_secure(request)
             response.set_cookie(
                 self.cookie_name, cookie, secure=secure, httponly=True